requirements.txt
uvicorn~=0.32.0
SQLAlchemy~=2.0.36
fastapi~=0.115.5
passlib~=1.7.4
jose~=1.0.0
pydantic~=2.9.2
pytest~=8.3.3
httpx~=0.27.2
PyJWT~=2.9.0
psycopg2-binary>=2.9.0
pydantic[email]
bcrypt
orjson
DateTime~=5.5
uuid~=1.30
python-dotenv~=1.0.1